Refactor asset package a bit, mainly just putting the etcd and postgres specific stuff in separate files
