+++ src/server/pkg/deploy/assets/util.go
@@ -1,0 +1,134 @@
+package assets
+
+import (
+	"fmt"
+	"path"
+	"strings"
+
+	"github.com/pachyderm/pachyderm/src/client/pkg/errors"
+	v1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+func makeStorageClass(opts *AssetOpts, backend Backend, storageClassName string, storageClassLabels map[string]string) (interface{}, error) {
+	sc := map[string]interface{}{
+		"apiVersion": "storage.k8s.io/v1",
+		"kind":       "StorageClass",
+		"metadata": map[string]interface{}{
+			"name":      storageClassName,
+			"labels":    storageClassLabels,
+			"namespace": opts.Namespace,
+		},
+		"allowVolumeExpansion": true,
+	}
+	switch backend {
+	case GoogleBackend:
+		sc["provisioner"] = "kubernetes.io/gce-pd"
+		sc["parameters"] = map[string]string{
+			"type": "pd-ssd",
+		}
+	case AmazonBackend:
+		sc["provisioner"] = "kubernetes.io/aws-ebs"
+		sc["parameters"] = map[string]string{
+			"type": "gp2",
+		}
+	default:
+		return nil, nil
+	}
+	return sc, nil
+}
+
+func makePersistentVolume(opts *AssetOpts, persistentDiskBackend Backend, hostPath string, name string, size int, volumeName string, volumeLabels map[string]string) (*v1.PersistentVolume, error) {
+	spec := &v1.PersistentVolume{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "PersistentVolume",
+			APIVersion: "v1",
+		},
+		ObjectMeta: objectMeta(volumeName, volumeLabels, nil, opts.Namespace),
+		Spec: v1.PersistentVolumeSpec{
+			Capacity: map[v1.ResourceName]resource.Quantity{
+				"storage": resource.MustParse(fmt.Sprintf("%vGi", size)),
+			},
+			AccessModes:                   []v1.PersistentVolumeAccessMode{v1.ReadWriteOnce},
+			PersistentVolumeReclaimPolicy: v1.PersistentVolumeReclaimRetain,
+		},
+	}
+
+	switch persistentDiskBackend {
+	case AmazonBackend:
+		spec.Spec.PersistentVolumeSource = v1.PersistentVolumeSource{
+			AWSElasticBlockStore: &v1.AWSElasticBlockStoreVolumeSource{
+				FSType:   "ext4",
+				VolumeID: name,
+			},
+		}
+	case GoogleBackend:
+		spec.Spec.PersistentVolumeSource = v1.PersistentVolumeSource{
+			GCEPersistentDisk: &v1.GCEPersistentDiskVolumeSource{
+				FSType: "ext4",
+				PDName: name,
+			},
+		}
+	case MicrosoftBackend:
+		dataDiskURI := name
+		split := strings.Split(name, "/")
+		diskName := split[len(split)-1]
+
+		spec.Spec.PersistentVolumeSource = v1.PersistentVolumeSource{
+			AzureDisk: &v1.AzureDiskVolumeSource{
+				DiskName:    diskName,
+				DataDiskURI: dataDiskURI,
+			},
+		}
+	case MinioBackend:
+		fallthrough
+	case LocalBackend:
+		pathType := v1.HostPathDirectoryOrCreate
+		spec.Spec.PersistentVolumeSource = v1.PersistentVolumeSource{
+			HostPath: &v1.HostPathVolumeSource{
+				Path: path.Join(hostPath, volumeName),
+				Type: &pathType,
+			},
+		}
+	default:
+		return nil, errors.Errorf("cannot generate volume spec for unknown backend \"%v\"", persistentDiskBackend)
+	}
+	return spec, nil
+}
+
+func makeVolumeClaim(opts *AssetOpts, size int, volumeName string, volumeClaimName string, volumeClaimLabels map[string]string) *v1.PersistentVolumeClaim {
+	return &v1.PersistentVolumeClaim{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "PersistentVolumeClaim",
+			APIVersion: "v1",
+		},
+		ObjectMeta: objectMeta(volumeClaimName, volumeClaimLabels, nil, opts.Namespace),
+		Spec: v1.PersistentVolumeClaimSpec{
+			Resources: v1.ResourceRequirements{
+				Requests: map[v1.ResourceName]resource.Quantity{
+					"storage": resource.MustParse(fmt.Sprintf("%vGi", size)),
+				},
+			},
+			AccessModes: []v1.PersistentVolumeAccessMode{v1.ReadWriteOnce},
+			VolumeName:  volumeName,
+		},
+	}
+}
+
+func makeHeadlessService(opts *AssetOpts, name, serviceName string, ports []v1.ServicePort) *v1.Service {
+	return &v1.Service{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Service",
+			APIVersion: "v1",
+		},
+		ObjectMeta: objectMeta(serviceName, labels(name), nil, opts.Namespace),
+		Spec: v1.ServiceSpec{
+			Selector: map[string]string{
+				"app": name,
+			},
+			ClusterIP: "None",
+			Ports:     ports,
+		},
+	}
+}
